refactored projects dialog from add package
